Moved policy-related rules into it's own set
